// backend/src/enums/user.enum.ts
export enum Status {
    Unverified = "UNVERIFIED",
    PendingVerification = "PENDING VERIFICATION",
    Verified = "VERIFIED",
}

export enum Gender {
    Male = "MALE",
    Female = "FEMALE",
}

export enum MaritalStatus {
    Single = "SINGLE",
    Married = "MARRIED",
    Divorced = "DIVORCED",
    Widowed = "WIDOWED"
}

export enum Role {
    Admin = "ADMIN",
    Normal = "NORMAL"
}

export enum Type {
    NID = "NID",
    Passport = "PASSPORT",
}